♻️ separate utils from telegram
